+++ lib/service_actor/result.rb
@@ -11,7 +11,18 @@ class ServiceActor::Result < BasicObject
     end
   end
 
-  %i[class is_a? kind_of? send public_send tap then block_given? object_id instance_variables].each do |method_name|
+  %i[
+    class
+    is_a?
+    kind_of?
+    send
+    public_send
+    tap
+    then
+    block_given?
+    object_id
+    instance_variables
+  ].each do |method_name|
     define_method(method_name, ::Kernel.instance_method(method_name))
   end
 
@@ -49,6 +60,10 @@ class ServiceActor::Result < BasicObject
 
   def failure?
     data[:failure] || false
+  end
+
+  def error
+    data[:error] || nil
   end
 
   def merge!(result)